+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[FIX]** Make `Paragraph` layout resilient to long strings without spaces.
+- **[FIX]** Fix `PetIcon` stroke fill color.
 [...]
 
 # v34.4.0 (18/06/2020)