+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Unreleased
 
+[...]
+
+# v34.4.1 (22/06/2020)
+
 - **[FIX]** Make `Paragraph` layout resilient to long strings without spaces.
 - **[FIX]** Fix `PetIcon` stroke fill color.
-[...]
 
 # v34.4.0 (18/06/2020)